c# updates

+++ Magnetar/Core/InternalCalls.cs
@@ -6,10 +6,10 @@ namespace Magnetar.Core
     internal static class InternalCalls
     {
         [MethodImpl(MethodImplOptions.InternalCall)]
-        public static extern T GetComponent<T>(ulong id);
+        public static extern T GetComponent<T>(uint id);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        public static extern bool HasComponent(ulong id, Type type);
+        public static extern bool HasComponent(uint id, Type type);
         
         [MethodImpl(MethodImplOptions.InternalCall)]
         public static extern void Log(LogLevel level, string message);
